Add YAML configuration file support
The sample YAML config files do not have %(here)s, or any other
interpolation.  The plaster-yaml interpolation code is broken, and, in
any case, should probably be removed and replaced with template
substitution.

The %(here)s are removed from testing.ini, in favor of a relative path.

The logging configuration in the testing config files has "%" changed
to "%%".  This is so the sqlalchemy conftest.py can copy configuration
and manually build the alembic configuration, completely copying
configuration without having to include or exclude particular
sections, and not have alembic complain about interpolation errors.
It seems that alembic has no way to turn off configuration
interpolation, and interpolation cannot be turned off in
configparser.Config() instances once they are created.

At the time of this writing there are a number of PRs/patches that
must be applied to dependent pacakges:

 * The Loader.get_settings() function does not have a "raw" parameter
   in any of: plaster, plaster_pastedeploy, plaster-yaml
 * plaster-yaml has a bug in Loader.get_sections().